remove slack logger worker
